+++ src/cmd/jpg/ico.c
@@ -47,7 +47,7 @@ gets(uchar *p)
 ulong
 getl(uchar *p)
 {
-	return p[0] | (p[1]<<8) | (p[2]<<16) | (p[3]<<24);
+	return p[0] | (p[1]<<8) | (p[2]<<16) | ((ulong)p[3]<<24);
 }
 
 int
